Move shortcuts modal to sidebar
